Add `unfold` builder
We don't want the default norm builder to add unfoldable definitions to
the simp set, because then it basically always applies (and usually does
the wrong thing). So now there's a separate `unfold` builder.

Todo: Maybe generalise the simp builder again so that it adds either an
equation or an unfoldable definition. Then the simp builder would be
equivalent to @[simp] on the definition. We might then want a simp_eq
builder that only adds equations.